fix(typeorm): :zap: improve migration generation
Make middle_name nullable.
The migration:generate command now checks if we already have unapplied migrations in the typeorm's migrations directory, if yes then it will show the unapplied migrations and abort the process otherwise it will inspect the db for changes with respect to the entities and generate the migration.